+++ js/function.js
@@ -57,3 +57,23 @@ getNumber('2023 год');
 //console.log(getNumber('-1'));
 //console.log(getNumber('1.5'));
 
+const strokeTimeToMinutes = (timeInArray) => timeInArray[0] * 60 + Number(timeInArray[1]);
+
+const meetingIsReally = (workingTimeStart, workingTimeEnd, meetingTimeStart, meetingDuration) => {
+  const workingTimeStartInMinutes = strokeTimeToMinutes(workingTimeStart.split(':'));
+  const workingTimeEndInMinutes = strokeTimeToMinutes(workingTimeEnd.split(':'));
+  const meetingTimeStartInMinutes = strokeTimeToMinutes(meetingTimeStart.split(':'));
+
+  if (workingTimeStartInMinutes <= meetingTimeStartInMinutes && workingTimeEndInMinutes >= (meetingTimeStartInMinutes + meetingDuration)) {
+    return true;
+  }
+  return false;
+};
+
+meetingIsReally('08:00', '17:30', '14:00', 90);
+
+// console.log(meetingIsReally('08:00', '17:30', '14:00', 90)); // true
+// console.log(meetingIsReally('8:0', '10:0', '8:0', 120));     // true
+// console.log(meetingIsReally('08:00', '14:30', '14:00', 90)); // false
+// console.log(meetingIsReally('14:00', '17:30', '08:0', 90));  // false
+// console.log(meetingIsReally('8:00', '17:30', '08:00', 900)); // false
